[#45] Implement time and place rescheduling logic
